Feat. Added all operate functions to main api
