+++ classes/api_main.py
@@ -147,4 +147,105 @@ class Api():
     
     @property
     def get_balance(self):
-        return self.API.get_balance()
+        return self.banca
+
+    def update_balance(self):
+        self.banca = self.API.get_balance()
+
+    def set_status(self):
+        self.status = not self.status
+
+    @property
+    def get_status(self):
+        return self.status
+
+    # Aguarda o momento correto para realizar a entrada.
+    def its_time(self, hour):
+        now = datetime.now()
+        #print(hour)
+        date = datetime(int(now.strftime("%Y")), int(now.strftime("%m")), int(now.strftime("%d")), int(hour[0]), int(hour[1]), 00)
+        print("Aguardando o horário de compra")
+        while True:
+            if not self.stop(): break
+            #print(date ,datetime.timestamp(date), self.API.get_server_timestamp())
+            if (datetime.timestamp(date)-2) >= self.API.get_server_timestamp() and datetime.timestamp(date) <= (self.API.get_server_timestamp()+5):
+                return True
+            if (datetime.timestamp(date)+10) < self.API.get_server_timestamp():
+                print("Horário de compra passou.")
+                return False
+
+    # Parâmetros de parada do automatizador.
+    def stop(self):
+        if self.status == True:
+            return False
+        if self.API.get_balance() >= (self.banca + self.stop_win):
+            print("Stop win batido.")
+            input("Pressione enter para sair.")
+            return False
+        if self.API.get_balance() <= (self.banca - self.stop_loss):
+            print("Stop loss batido.")
+            input("Pressione enter para sair.")
+            return False
+        if self.API.get_balance() < (self.banca + self.stop_win) and self.API.get_balance() > (self.banca - self.stop_loss): return True
+  
+    # Abre as ordens em opções binárias.
+    def buy_binary(self, active, action, duration, time):
+        value = self.value
+        for trie in range(self.martingale):
+            check,id=self.API.buy(value, active, action, duration)
+            if check:
+                if value > self.__value: print(f"{trie}ª Martingale")
+                print(f"Ordem de {action} em {active} de {value} aberta.")
+            else:
+                print(f"Oops, tive um problema em abrir a ordem {active} as {time}")
+                return 0
+            result = self.API.check_win_v3(id)
+            if result < 0: 
+                if trie == self.martingale-1:
+                    self.txt.write(f'''R$ {result} - {active} - {time} - {action}''')
+                value*= self.multiplicador
+            print(f"Resultado da ordem {result}")
+            if result > 0: 
+                self.txt.write(f'''R$ +{result} - {active} - {time} - {action}''')
+                return result
+
+    # Abre as ordens em opções digital.
+    def buy_digital(self, active, action, duration, time):
+        value = self.value
+        for trie in range(self.martingale):
+            _, id = self.API.buy_digital_spot(active, value, action.lower(), duration)
+            if id !="error":
+                if value > self.value : print(f"{trie}ª Martingale")
+                print(f"Ordem de {action} em {active} de {value} aberta.")
+                while True:
+                    check,win=self.API.check_win_digital_v2(id)
+                    if check==True:
+                        break
+                if win<0:
+                    print(f"Resultado da ordem -R${win}")
+                    value *= self.multiplicador
+                    if trie == self.martingale-1:
+                        self.txt.write(f'''R$ -{win} - {active} - {time} - {action}''')
+                if win > 0:
+                    print(f"Resultado da ordem +R${win}")
+
+                    self.txt.write(f'''R$ +{win} - {active} - {time} - {action}''')
+                    return win
+            else:
+                print(f"Oops, tive um problema em abrir a ordem {active} as {time}")
+
+    # Inicia as operações.
+    def operate(self):
+        self.status = False
+        for item in self.sinais:
+            print(item)
+            if self.stop():
+                entrar = self.its_time(item['Hour'])
+                if entrar == True:
+                    if self.option == "BINARY":
+                        self.buy_binary(item["Active"], item["Action"], item["Duration"], item["Hour"])
+                    else:    
+                        self.buy_digital(item["Active"], item["Action"], item["Duration"], item["Hour"])
+                else: continue
+
+        print("acabou")
